+++ pkg/auth/types.go
@@ -12,9 +12,9 @@ type LoginInfo struct {
 }
 
 type Tokens struct {
-	TokenType    string  `json:"token_type"`
-	ExpiresIn    float64 `json:"expires_in"`
-	Scope        string  `json:"scope"`
-	AccessToken  string  `json:"access_token"`
-	RefreshToken string  `json:"refresh_token"`
+	TokenType    string `json:"token_type"`
+	ExpiresIn    int    `json:"expires_in"`
+	Scope        string `json:"scope"`
+	AccessToken  string `json:"access_token"`
+	RefreshToken string `json:"refresh_token"`
 }
